Update MyRadioService.java
This single addition of .setForegroundServiceType(NotificationCompat.FOREGROUND_SERVICE_TYPE_MEDIA_PLAYBACK) tells the Android system that the service is a media playback service, which is essential when the service is intended to perform operations that should be noticeable by the user at all times (like playing audio). This change will ensure compliance with Android's foreground service requirements and prevent the app from crashing due to missing service type specifications.